init: Added parsing of symbolic GPIOs in bootloader sequence
There are some kernels in which the sysfs interface GPIOs are exported
with filenames differing from the standard "gpio%d" (e.g.: pioe11).
Allow using those "symbolic GPIOs" in the bootloader sequence in
addition to being able to still use the integers way.

Since GPIOs can not be exported with their symbolic name, the program
will fail if some unexported symbolic GPIO is used in the bootloader
sequence.